+++ src/main.rs
@@ -1,112 +1,10 @@
-use serde::de::{DeserializeOwned, Error};
-use serde::{Deserialize, Serialize};
-use std::fs::File;
-use std::io::Read;
-use std::path::{Path, PathBuf};
-use chrono::Local;
+mod settings;
+use std::path::PathBuf;
 
-#[derive(Debug, Deserialize, Serialize, PartialEq)]
-struct Settings {
-    save_base_path: PathBuf,
-    game_conf_path: PathBuf,
-    color_scheme: String,
-    delete_on_restore: bool,
-}
+use settings::Settings;
+mod config;
+use config::{game::Game, verify_conf, write_conf};
 
-#[allow(dead_code)]
-#[derive(Debug, Serialize, Deserialize)]
-struct Game {
-    game_title: String,
-    steam_id: u32,
-    save_path: PathBuf,
-    publisher: String,
-    developer: String,
-    saves: Vec<Save>,
-}
-impl Game {
-    /**
-    Adds a season to the Show.
-
-    # Example
-    ```
-    let mut er = Game { game_title: "Elden Ring".to_string(), steam_id: 1245620, save_path: PathBuf::from("/mnt/storage/SteamLibrary/steamapps/compatdata/1245620/"), publisher: "Bandai Namco".to_string(), developer: "FROM Software".to_string(), Saves: vec![] };
-    let path:String = format!("{}/{}",base_path, er.game_title);
-    er.add_save(path);
-    ```
-    # This adds the save to the game, to later make the backup.
-    */
-    fn add_save(&mut self, production_path: PathBuf, settings_path: &Path) {
-        // NOTE Is this the most efficient manner to get the count?
-        let count = self
-            .saves
-            .iter()
-            .max_by_key(|save| save.count)
-            .map(|save| save.count + 1)
-            .unwrap_or(0);
-        // NOTE parent_game: helps backup_path
-        let parent_game = self.game_title.clone();
-        // NOTE  backup_path: simply a path made up of the path defined in your settings, the name of the game, and the count of the settings.
-        let backup_path: PathBuf = PathBuf::from(format!("{}/{}/{}", settings_path.to_str().unwrap_or("/home/user/"), &parent_game, &count));
-        // NOTE Should this be in epoch and converted later with a TZ defined by the user, or should it be converted now?
-        let saved_at = Local::now().naive_local().format("%Y-%m-%dT%H:%M:%SZ").to_string();
-        let new_save: Save = Save {
-        count,
-        backup_path,
-        production_path,
-        parent_game,
-        saved_at,
-    };
-        self.saves.push(new_save);
-    }
-}
-#[allow(dead_code)]
-#[derive(Debug, Serialize, Deserialize)]
-struct Save {
-    count: u8,
-    backup_path: PathBuf,
-    production_path: PathBuf,
-    parent_game: String,
-    saved_at: String,
-}
-fn write_conf<T>(conf: Vec<T>, pth: &Path)
-// fn write_conf<T>(conf: Vec<T>, pth: PathBuf -> serde_json::Result<()>) 
-where
-    T: Serialize,
-{
-    if let Ok(out_file) = File::create(pth){
-        if let Err(err) = serde_json::to_writer(out_file, &conf){
-            eprintln!("Oh noes: {}",err);
-        }
-    }else if let Err(err) = File::create(pth) {
-        eprintln!("Error creating file {:?}: {}", pth, err);
-    }
-}
-
-fn read_conf<T>(conf_data: PathBuf) -> Result<T, serde_json::Error>
-where
-    T: DeserializeOwned,
-{
-    let mut file = File::open(conf_data).map_err(|e| serde_json::Error::custom(e.to_string()))?;
-    let mut json_data = String::new();
-    file.read_to_string(&mut json_data)
-        .map_err(|e| serde_json::Error::custom(e.to_string()))?;
-    let settings: T = serde_json::from_str(&json_data)?;
-    Ok(settings)
-}
-
-// Define a function to verify and return the configuration
-fn verify_conf<T>(conf_path: PathBuf) -> T
-where
-    T: DeserializeOwned,
-{
-    match read_conf(conf_path) {
-        Ok(conf) => conf,
-        Err(err) => {
-            eprintln!("Error: {}", err);
-            std::process::exit(1);
-        }
-    }
-}
 
 fn main() {
     // Get the user's home directory
@@ -130,9 +28,10 @@ fn main() {
 // Ai Tests
 #[cfg(test)]
 mod tests {
+    use std::io::Write;
+    use crate::config::read_conf;
     use super::*;
     use assert_matches::assert_matches;
-    use std::io::Write;
 
     #[test]
     fn fuzz_bad_settings() {
